requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
awpy==2.0.0b4
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
demoparser2==0.34.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
fonttools==4.55.3
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.8
loguru==0.7.3
matplotlib==3.10.0
multidict==6.1.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==10.4.0
polars==1.19.0
propcache==0.2.1
pyarrow==18.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
scipy==1.15.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
usd-core==24.11
win32_setctime==1.2.0
yarl==1.18.3